[Feature][Refactor] Refactored Engine Core + Editor Developments (#52)

+++ Scribe/Source/Panels/SceneHierarchyPanel.cpp
@@ -19,23 +19,21 @@ namespace Scribe {
 
             scene->component<Sentinel::TransformComponent>();
             scene->each<Sentinel::TransformComponent>([&](flecs::entity e, Sentinel::TransformComponent& transform) {
-                if (!m_LoadedScene->registry[e]->HasParent()) { DisplayNode(m_LoadedScene->registry[e]); }
+                Sentinel::Entity entity(e, m_LoadedScene);
+                if (!entity.HasParent()) { DisplayNode(entity); }
             });
 
             ImGui::End();
 
-            if (m_MarkedForDelete != nullptr) {
-                m_LoadedScene->DeleteEntity(m_MarkedForDelete);
-                m_MarkedForDelete = nullptr;
-            }
+            if (m_MarkedForDelete.GetNative().is_valid()) { m_LoadedScene->DeleteEntity(m_MarkedForDelete); }
         }
 
-        void SceneHierarchyPanel::DisplayNode(Sentinel::Entity* entity) {
+        void SceneHierarchyPanel::DisplayNode(Sentinel::Entity entity) {
             ImGuiTreeNodeFlags flags = ImGuiTreeNodeFlags_SpanAvailWidth;
-            flags |= entity->HasChildren() ? ImGuiTreeNodeFlags_OpenOnArrow : ImGuiTreeNodeFlags_Leaf;
+            flags |= entity.HasChildren() ? ImGuiTreeNodeFlags_OpenOnArrow : ImGuiTreeNodeFlags_Leaf;
 
-            flecs::entity* e = entity->GetNative();
-            bool opened = ImGui::TreeNodeEx((void*)(Sentinel::UInt64)*e, flags, e->name().c_str());
+            flecs::entity e = entity.GetNative();
+            bool opened = ImGui::TreeNodeEx((void*)(Sentinel::UInt64)e, flags, e.name().c_str());
             // if (!opened && ImGui::IsItemClicked(0)) m_SelectedEntity = entity;
 
             if (ImGui::BeginPopupContextItem()) {
@@ -44,8 +42,9 @@ namespace Scribe {
             }
 
             if (opened) {
-                e->children([&](flecs::entity e) {
-                    DisplayNode(m_LoadedScene->registry[e]);
+                e.children([&](flecs::entity e) {
+                    Sentinel::Entity childEntity(e, m_LoadedScene);
+                    DisplayNode(childEntity);
                 });
                 ImGui::TreePop();
             }
